ai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sentence-transformers==3.2.1
numpy==2.1.1
scikit-learn==1.5.2
mysql-connector-python==9.1.0
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2